some minor things.
got rid of bools in plfs.h so it can build with stupid C

changed it so comparisons of a void* and an int upgrade the int
to a void* instead of downgrading the void* to an int since on
64 bit machines, the compiler won't let us downgrade a void* to an int

fixed an ugly mmap leak



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/trunk@379 979fb63a-54b4-4514-8d67-9ab8ea1b59df

+++ tools/plfs_map.C
@@ -6,7 +6,7 @@
 int main (int argc, char **argv) {
     const char *target  = argv[1];
     bool compress = true;
-    int ret = plfs_dump_index(stderr,target,compress);
+    int ret = plfs_dump_index(stderr,target,(int)compress);
     if ( ret != 0 ) {
         fprintf(stderr, "Couldn't read index from %s: %s\n",
                 target, strerror(-ret));
